Refactor: split App.tsx into LanguageSelector and ChatView components
